moving ButtonVariant to Button's file
