--- src/@types/ButtonVariant.ts
@@ -1,6 +1,0 @@
-export enum ButtonVariant {
-  PRIMARY = "PRIMARY",
-  SECONDARY = "SECONDARY",
-  DANGER = "DANGER",
-  GHOST = "GHOST",
-}
